+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Extend unit tests for `SymbolTable` 
 - Rename `SymbolTableWriter` to `SymbolListingWriter` and make it stateless
 - Make AssemblerEngine stateless
 - Project structure: folders, CMake targets and modules reorganised